+++ tests/unit/libraries/Jentil/Setups/Translations/CoreTest.php
@@ -16,21 +16,14 @@ class CoreTest extends AbstractTestCase
     {
         $add_action = FunctionMocker::replace('add_action');
 
-        $jentil = new class extends AbstractTheme {
-            function __construct()
-            {
+        $jentil = Stub::makeEmpty(AbstractTheme::class, [
+            'theme' => new class {
+                function get()
+                {
+                    return 'jentil';
+                }
             }
-
-            function get()
-            {
-                return new class {
-                    function get()
-                    {
-                        return 'jentil';
-                    }
-                };
-            }
-        };
+        ]);
 
         $translation = new Core($jentil);
 
@@ -47,23 +40,16 @@ class CoreTest extends AbstractTestCase
     {
         $load = FunctionMocker::replace('load_theme_textdomain');
 
-        $jentil = new class extends AbstractTheme {
-            function __construct()
-            {
+        $jentil = Stub::makeEmpty(AbstractTheme::class, [
+            'utilities' => Stub::makeEmpty(Utilities::class),
+            'theme' => new class {
+                function get()
+                {
+                    return 'jentil';
+                }
             }
+        ]);
 
-            function get()
-            {
-                return new class {
-                    function get()
-                    {
-                        return 'jentil';
-                    }
-                };
-            }
-        };
-
-        $jentil->utilities = Stub::makeEmpty(Utilities::class);
         $jentil->utilities->fileSystem = Stub::makeEmpty(FileSystem::class, [
             'dir' => '/var/www/jentil/languages',
         ]);